requirements.txt
# Django
Django==4.2.18
django-environ==0.11.2

# Base de données
psycopg2-binary==2.9.9

# Gestion des images
Pillow==10.1.0

# Reconnaissance faciale et Vision par ordinateur
opencv-python==4.9.0.80

numpy==1.26.2
mediapipe==0.10.21

# Utilitaires
python-dateutil==2.8.2
python-dotenv==1.0.0

# Développement
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0